api/requirements.txt
fastapi
uvicorn
cassandra-driver
